Refactor MCP
